Refactor as locals.js
