+++ config/index.js
@@ -20,19 +20,4 @@ module.exports = function () {
   app.use(errorHandler());
   // Add static path
   app.use(express.static(path.join(rootDirName, 'public')));
-
-  app.locals.clients = {};
-
-  const NorthstarClient = require('@dosomething/northstar-js');
-  app.locals.clients.northstar = new NorthstarClient({
-    baseURI: process.env.DS_NORTHSTAR_API_BASEURI,
-    apiKey: process.env.DS_NORTHSTAR_API_KEY,
-  });
-
-  const PhoenixClient = require('@dosomething/phoenix-js');
-  app.locals.clients.phoenix = new PhoenixClient({
-    baseURI: process.env.DS_PHOENIX_API_BASEURI,
-    username: process.env.DS_PHOENIX_API_USERNAME,
-    password: process.env.DS_PHOENIX_API_PASSWORD,
-  });
 };
